requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # for form-data (Twilio webhooks)

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0                  # async PostgreSQL driver
geoalchemy2==0.14.6              # PostGIS Geography types for SQLAlchemy
psycopg2-binary==2.9.9           # sync driver needed by Alembic CLI

# ─── Migrations ───────────────────────────────────────────────────────────────
alembic==1.13.1

# ─── Config & Validation ──────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.27.0

# ─── Comms ────────────────────────────────────────────────────────────────────
twilio==9.0.5

# ─── Azure SDKs (Step 2+) ─────────────────────────────────────────────────────
azure-ai-translation-text==1.0.0b1   # Azure AI Translator
azure-ai-contentsafety==1.0.0        # Azure AI Content Safety

# ─── Azure SDKs (Step 3+) ─────────────────────────────────────────────────────
openai==1.30.1                       # Azure OpenAI (Planner Agent)
pyautogen==0.2.35                    # Microsoft AutoGen multi-agent framework
# azure-search-documents==11.4.0    # Azure AI Search (future RAG upgrade)
